+++ doc/release-notes.md
@@ -10,6 +10,7 @@
   - Blank/unset for active coins.
   - All for all known coins.
   - Comma separated list of coin tickers to show.
+- basicswap-run will rewrite litecoin.conf file to add config required to run Litecoin 0.21.3 in pruned mode.
 
 
 0.12.7